Change `restockCycles` to short size

+++ api/cache/src/main/kotlin/org/rsmod/api/cache/types/inv/InvTypeEncoder.kt
@@ -98,7 +98,7 @@ public object InvTypeEncoder {
                         data.writeInt(stock.count)
                     }
                     data.writeShort(stock.obj)
-                    data.writeByte(stock.restockCycles)
+                    data.writeShort(stock.restockCycles)
                 }
             }
         }
